move stats endpoints to new file
(no new tests for new endpoint included in this commit)
